% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graduation.R
\name{graduation}
\alias{graduation}
\title{graduation}
\usage{
graduation(data_select, data_ultimate, data_vbt)
}
\arguments{
\item{data_select}{raw mortality table in select period}

\item{data_ultimate}{raw mortality table in ultimate period}

\item{data_vbt}{standard mortality table}
}
\value{
graduated mortality table
}
\description{
This function loads raw mortality tables in select and ultimate periods,
and then do the graudation based on the trends in the standard table.
Finally output graduated mortality table
}
