% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_improvement.R
\name{mortality_improvement}
\alias{mortality_improvement}
\title{mortality_improvement}
\usage{
mortality_improvement(claims, vbt_year, vbt_month, factor)
}
\arguments{
\item{claims}{claims table with mortality actual amount larger than zero}

\item{vbt_year}{The year of vbt table}

\item{factor}{The factor is applied to the mortality improvement rates}

\item{data_vbt}{The month of vbt table}
}
\value{
claims table with calculated improvement factors
}
\description{
This function loads claims table with mortality actual amount larger than zero,
and then calculate improvement factors for each of claims.
Finally claims table with calculated improvement factors.
}
