% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_transformation.R
\name{table_transformation}
\alias{table_transformation}
\title{table_transformation}
\usage{
table_transformation(table)
}
\arguments{
\item{table}{mortality table downloaded from SOA website}
}
\value{
long mortality table ready to analyze
}
\description{
This function loads mortality table downloaded from SOA website,
and then transfer it to a long table ready to analyze
}
