% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wear_off.R
\name{wear_off}
\alias{wear_off}
\title{wear_off}
\usage{
wear_off(table, age_basis, gender, tobacco_status, insure_age)
}
\arguments{
\item{table}{transformed mortality table}

\item{age_basis}{age basis which can be "ANB" or "ALB"}

\item{gender}{gender which can be "Male" or "Female"}

\item{tobacco_status}{tobacco_status which can be "Nontobacco" or "Tobacco"}

\item{insure_age}{insure_age which can be from 18 to 95}
}
\value{
an vector of wear off mortality rates
}
\description{
This function loads transformed mortality table,
and then find out the wear off sequence of mortality rates given an insure age,
an age_basis, a gender, and a tobacco_status
}
